chore: move nlpService into packages/client/src/services
