+++ packages/client/src/client/OpenFin/apps/Launcher/NlpSuggestions/SuggestionWrapper.tsx
@@ -2,9 +2,10 @@ import { Subscribe } from "@react-rxjs/core"
 import { Loader } from "client/components/Loader"
 import { WithChildren } from "client/utils/utilityTypes"
 
+import { NlpIntent } from "@/services/nlp"
+
 import { appConfigs } from "../applicationConfigurations"
 import { reactiveTraderFxIcon } from "../icons"
-import { NlpIntent } from "../services/nlpService"
 import { open } from "../tools"
 import { handleIntent } from "./intents"
 import {
