+++ packages/client/src/client/OpenFin/apps/Launcher/NlpSuggestions/Trades.tsx
@@ -2,7 +2,8 @@ import { formatNumber } from "client/utils"
 import { format } from "date-fns"
 import { useTrades } from "services/trades"
 
-import { TradesInfoIntent } from "../services/nlpService"
+import { TradesInfoIntent } from "@/services/nlp"
+
 import {
   Col,
   ResultsTable,
