+++ packages/client/src/services/nlp/index.ts
@@ -1,0 +1,1 @@
+export * from "./nlp"
